[#30] hrタグをBlock要素へと変更したことに伴い、テストコードも見直し
