+++ tests/markdown/inline_parser_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from app.markdown.inline_parser import InlineParser, LinkParser, CodeParser, ImageParser, HorizontalRuleParser
+from app.markdown.inline_parser import InlineParser, LinkParser, CodeParser, ImageParser
 
 
 class TestInlineParser:
@@ -212,58 +212,6 @@ class TestImage:
     def test_parse(self, text: str, expected: str):
         # GIVEN
         sut = ImageParser()
-        # WHEN
-        actual = sut.parse(text)
-        # THEN
-        assert repr(actual) == expected
-
-
-class TestHorizontalRule:
-    """ ---で表現される水平線罫線要素を検証 """
-
-    # 記法が対象か
-    @pytest.mark.parametrize(
-        ('text', 'expected'),
-        [
-            ('---', True),
-            ('罫線のつもりです', False),
-        ],
-        ids=['hr', 'not hr'])
-    def test_target(self, text: str, expected: bool):
-        # GIVEN
-        sut = HorizontalRuleParser()
-        # WHEN
-        actual = sut.is_target(text)
-        # THEN
-        assert actual == expected
-
-    # 記法に基づいて分離
-    @pytest.mark.parametrize(
-        ('text', 'expected'),
-        [
-            (
-                    '---',
-                    ('', '---', '')
-            ),
-        ])
-    def test_extract(self, text: str, expected: tuple[str, str, str]):
-        # GIVEN
-        sut = HorizontalRuleParser()
-        # WHEN
-        actual = sut.extract_text(text)
-        # THEN
-        assert actual == expected
-
-    # 記法を解釈
-    @pytest.mark.parametrize(
-        ('text', 'expected'),
-        [
-            ('---', 'HorizontalRule: ')
-        ]
-    )
-    def test_parse(self, text: str, expected: str):
-        # GIVEN
-        sut = HorizontalRuleParser()
         # WHEN
         actual = sut.parse(text)
         # THEN
