+++ tests/template/html/block.html
@@ -52,9 +52,7 @@
         const text = 'Hello Block'
     </code>
 </pre>
-<p class="mt-2 mb-2">
-    
-</p>
+<hr class="border-b-2 border-indigo-400">
 <h2 class="pb-1 mt-4 mb-4 font-sans text-2xl font-semibold border-b-2 border-indigo-400 border-solid">
     終わり
 </h2>
