+++ tests/template/html/inline.html
@@ -19,11 +19,5 @@
     
 </p>
 <p class="mt-2 mb-2">
-    <hr class="border-b-2 border-indigo-400">
-</p>
-<p class="mt-2 mb-2">
-    
-</p>
-<p class="mt-2 mb-2">
     <img src="https://avatars.githubusercontent.com/u/43694794?v=4" alt="いい感じ">
 </p>
