+++ tests/template/html/sample_article.html
@@ -83,9 +83,7 @@
 <p class="mt-2 mb-2">
     
 </p>
-<p class="mt-2 mb-2">
-    <hr class="border-b-2 border-indigo-400">
-</p>
+<hr class="border-b-2 border-indigo-400">
 <p class="mt-2 mb-2">
     
 </p>
