+++ tests/template/markdown/block.md
@@ -16,5 +16,5 @@
 # comment
 const text = 'Hello Block'
 ```
-
+---
 ## 終わり
