+++ tests/template/markdown/inline.md
@@ -6,6 +6,4 @@
 コードの中では[link](url)も無効化されます
 ```
 
----
-
 ![いい感じ](https://avatars.githubusercontent.com/u/43694794?v=4)
